fix GDScript and GDExtension binding errors for Godot 4.7
- Remove manual _process bind_method (auto-bound by GDCLASS virtuals)
- Bind property setters/getters before ADD_PROPERTY in _bind_methods
- Change _process(float) to _process(double) to match base class
- Remove stale .gdns preloads from import_midi/soundfont_plugin.gd
- Rename load() -> _load() with 4 params, save() -> _save() with return
- Add null guard on FileAccess.open() in resource loaders
- Replace .gdns ExtResource in main.tscn with GDMidiAudioStreamPlayer type
- Add root-level .gdextension file

+++ src/gdmidiplayer.h
@@ -41,7 +41,7 @@ public:
 
 	void _init(); // our initializer called by Godot
 
-	void _process(float delta);
+	void _process(double delta);
 	void set_soundfont(String p_soundfont);
 	String get_soundfont();
